+++ package.json
@@ -51,6 +51,7 @@
     "@types/flatbush": "^3.1.0",
     "@types/jimp": "^0.2.28",
     "@types/minimist": "^1.2.0",
+    "@types/ndjson": "^2.0.0",
     "@types/node": "12",
     "@types/shortid": "^0.0.29",
     "@types/tmp": "^0.1.0",
@@ -59,11 +60,14 @@
     "electron-webpack": "^2.8.2",
     "electron-webpack-ts": "^4.0.1",
     "electron-window-state": "^5.0.3",
-    "husky": "^4.2.5",
-    "lint-staged": "^10.2.0",
-    "prettier": "^2.0.5",
-    "typescript": "^3.8.3",
-    "webpack": "~4.42.1"
+    "husky": "^4.3.0",
+    "lint-staged": "^10.4.0",
+    "meow": "^7.1.1",
+    "ndjson": "^2.0.0",
+    "prettier": "^2.1.2",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3",
+    "webpack": "~4.44.2"
   },
   "prettier": {
     "singleQuote": false,
